+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - A fix for watching config file on linux systems
 - New `browser` option in the config file to indicate which browser should be launched.
 - Cache globs now match hostname + pathname, not just pathname
+- Logfiles are now one per day instead a single giant file
+- Fix unit tests, unskip disabled tests
 
 ## 0.0.8 Bugfixes, config watchmode